Lots of Command Stuff Many :bug:s to come
Added command to pluggedIn each command array has a name an array of
aliases a list of suggested args and  a callback function.

I then added a API.on(API.CHAT_COMMAND) that will sort through all the
pluggedIn.commands and their aliases and attempt to find a match.

This has minimal testing only on a command line level.  So many stupid
bugs like missing semicolons and missing parenthesis and syntax and just
plain out logical errors.   So hopefully this will work out.  (This is a
long one thanks for reading it all even though it is complete bullshit)